Fix the correct availability of sub-generators
Summary:
It could be possible that the sub-generators were not available because
the `package.json` was incomplete (it didn't include those folders with
the sub-generators).

This change comes from reading the documentation (http://yeoman.io/authoring)
and by checking the differences with `generator-angular`, which it does
have in the `package.json` listed its sub-generators.

Test Plan:
It can't be tested right until we deploy a new version and test
with a fresh installation of the generator.

Reviewers: jmsotuyo

Reviewed By: jmsotuyo

Differential Revision: http://ph.monits.com/D14985